Worked on PID and motor control

+++ Embedded/RMDBA_Motor_Control/Core/Inc/pid.h
@@ -22,38 +22,38 @@
   */
 typedef struct PID_Controller {
 	// Controller gains
-	double kp;	// Proportional gain
-	double ki;	// Integral gain
-	double kd;	// Derivative gain
+	float kp;	// Proportional gain
+	float ki;	// Integral gain
+	float kd;	// Derivative gain
 
 	// Discrete sampling time
-	double ts;
+	float ts;
 
 	// Controller hard limits (assumed to be symmetric)
-	double i_max;	// Maximum allowable error integral
-	double u_max;	// Maximum allowable control output
+	float i_max;	// Maximum allowable error integral
+	float u_max;	// Maximum allowable control output
 
 	// Derivative reference weighting (if not in use, set to 1)
-	double d_ref_wt;
+	float d_ref_wt;
 
 	// Dynamic values
-	double prev_ref;	// Prevoius reference
-	double prev_meas;	// Previous measurement
-	double i_sum;		// Error integral
+	float prev_ref;	// Prevoius reference
+	float prev_meas;	// Previous measurement
+	float i_sum;		// Error integral
 
 	// Controller output
-	double u;
+	float u;
 } PID_Controller;
 
 // Function prototypes
 /**
   * @brief	PID controller initialization
   */
-void pid_init(PID_Controller* cntlr, double kp, double ki, double kd, double ts, double i_max, double u_max, double d_ref_wt);
+void pid_init(PID_Controller* cntlr, float kp, float ki, float kd, float ts, float i_max, float u_max, float d_ref_wt);
 
 /**
   * @brief	PID controller output calculation
   */
-void pid_update(PID_Controller* cntlr, double ref, double meas);
+void pid_update(PID_Controller* cntlr, float ref, float meas);
 
 #endif /* INC_PID_H_ */
